Fixed a bug and added the import functionality.
The bug was in the end trial of a patient. A user could cancel without
selecting a reason. Now they must select a reason.

Importer is working super fly but, need to hook-up the UI and progress
for it and implement the states.

That just means only the Exporter is left for me to do :-D

Signed-off-by: Vincent J. Palodichuk <vj.palodichuk@gmail.com>